Major changes. Information in the description
checkValue:
Expanded the function a lot to validate the license correctly;

lockReservation:
Fixed a major bug where, if a user had locked a reservation and hadn't finished it yet, and another user decided to create a reservation on that timeslot, the second user wouldn't lock a new reservation and would instead overwrite the first user's reservation;

checkIfGroupDecrease:
Reworked the function to decrease each group's number that is bigger than the delete group's number by 1. The way it previously worked resulted in a bug that caused only the last group and the second to last group to decrease in number, and resulted in wrong numbering for other groups and timeslots overlapping.